Add files via upload
This is the project about cost of living across the globe. The dataset is sourced from Numbeo. There were 56 columns in the dataset including cities, countries and other aspects of daily life. 

I cleaned and transformed the dataset with the help of Power Query Editor and MS Excel. Then I created measures to get the comprehensive dataset. In the end I created the dashboard to get the major insights from the dataset.